RDKCOM-2351: RDKBDEV-595: make get_process_name() static
Reason for change:
move declaration of the caps variable from cap.h to cap.c + make static
Declaring the variable in cap.h without extern causes it to be
instantiated in each .c file which includes cap.h.
The normal fix would be to move the variable to cap.c and keep an
extern delaration in cap.h. However since the variable seems to be
private to cap.c, just move it there and make static.
The get_process_name() function is not a public API and so should not
be declared in cap.h. It was added recently as part of a change to
fix compiler warnings.
Test Procedure: Sanity.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit 80d8fe52547b208ab7e9119f271ca5ca5a9331a6)
Change-Id: Iab1be1403ab36a8f070948c00d0ac3cade575899