move declaration of the caps variable from cap.h to cap.c + make static
Declaring the variable in cap.h without extern causes it to be
instantiated in each .c file which includes cap.h.

The normal fix would be to move the variable to cap.c and keep an
extern delaration in cap.h. However since the variable seems to be
private to cap.c, just move it there and make static.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>